chore: scope queries
